+++ src/main/java/controller/HttpMethod.java
@@ -1,0 +1,6 @@
+package controller;
+
+public enum HttpMethod {
+    GET,
+    POST
+}
