+++ src/main/java/controller/Route.java
@@ -1,0 +1,13 @@
+package controller;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Route {
+    String uri();
+    HttpMethod method() default HttpMethod.GET;
+}
